src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/selections-in-shadow-expected.txt

CONSOLE ERROR: line 40: Uncaught TypeError: host.createShadowRoot is not a function







FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
